pyproject.toml
[project]
name = "ai-tools-show"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "itsdangerous>=2.2.0",
    "langfuse>=2.60.1",
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
    "sqladmin>=0.20.1",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
]
